various fixes: move examples up, allow Path as input
