+++ webknossos/lint.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -eEuo pipefail
+
+pylint -j4 webknossos
+pylint -j4 examples/*.py
