+++ webknossos/typecheck.sh
@@ -1,7 +1,11 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -eEuo pipefail
-echo "Typecheck webknossos module:"
+
+echo -n "Typecheck webknossos: "
 python -m mypy -p webknossos --disallow-untyped-defs --show-error-codes --strict-equality --namespace-packages
 
-echo "Typecheck tests:"
+echo -n "Typecheck tests:      "
 python -m mypy -p tests --disallow-untyped-defs --show-error-codes --strict-equality --namespace-packages
+
+echo -n "Typecheck examples:   "
+python -m mypy -p examples --disallow-untyped-defs --show-error-codes --strict-equality --namespace-packages
