+++ webknossos/webknossos/skeleton/__init__.py
@@ -1,5 +1,7 @@
+from os import PathLike
+
 from webknossos.skeleton.entities import Graph, Group, Node, Skeleton
 
 
-def open_nml(file_path: str) -> Skeleton:
+def open_nml(file_path: PathLike) -> Skeleton:
     return Skeleton.from_path(file_path)
